feat: Update app metadata and package name
Refactors the application to use the `spese-prova` package name and updates its metadata. This includes changing the application name in `metadata.json` and adjusting the `package.json` file.

Additionally, this commit introduces a confirmation modal for closing the expense form if changes have been made. It also replaces an icon in the dashboard component and removes the `ForgotEmailScreen` from the authentication flow, redirecting to the forgot password screen instead. The phone number field in the API logic is now initialized as an empty string for consistency.